organize file and separate widget
