+++ lib/ui/pages/home_page.dart
@@ -1,0 +1,102 @@
+part of 'pages.dart';
+
+class MyHomePage extends StatefulWidget {
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Color(0xff4963CC),
+      body: ListView(
+        children: <Widget>[
+          Padding(
+            padding: EdgeInsets.only(top: 15.0, bottom: 10.0),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                IconButton(
+                  icon: Icon(Icons.arrow_back_ios, color: Colors.white),
+                  onPressed: () {},
+                ),
+                Container(
+                  width: 125.0,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: <Widget>[
+                      IconButton(
+                        icon: Icon(Icons.filter_list),
+                        color: Colors.white,
+                        onPressed: () {},
+                      ),
+                      IconButton(
+                        icon: Icon(Icons.menu),
+                        color: Colors.white,
+                        onPressed: () {},
+                      ),
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+          SizedBox(height: 25.0),
+          Padding(
+            padding: EdgeInsets.only(left: 40.0),
+            child: Row(
+              children: <Widget>[
+                Text(
+                  'Healthy',
+                  style: kTitleApp,
+                ),
+                SizedBox(width: 10.0),
+                Text(
+                  'Food',
+                  style: kTitleApp,
+                ),
+              ],
+            ),
+          ),
+          SizedBox(height: 40.0),
+          Container(
+            height: MediaQuery.of(context).size.height - 185.0,
+            decoration: BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.only(topLeft: Radius.circular(75.0)),
+            ),
+            child: ListView(
+              primary: false,
+              padding: EdgeInsets.only(left: 25.0, right: 20.0),
+              children: <Widget>[
+                Padding(
+                  padding: EdgeInsets.only(top: 45.0),
+                  child: Container(
+                    height: MediaQuery.of(context).size.height - 300.0,
+                    child: ListView(
+                      children: <Widget>[
+                        ItemFood(
+                            'assets/plate1.png', 'Salmon Bowl', 'Rp.40.000'),
+                        ItemFood(
+                            'assets/plate2.png', 'Spring Bowl', 'Rp.30.000'),
+                        ItemFood(
+                            'assets/plate3.png', 'Steak Bowl', 'Rp.70.000'),
+                        ItemFood(
+                            'assets/plate4.png', 'Mie Bowl', 'Rp.20.000'),
+                        ItemFood(
+                            'assets/plate5.png', 'Tuna Bowl', 'Rp.100.000'),
+                        ItemFood(
+                            'assets/plate6.png', 'Kakap Bowl', 'Rp.150.000'),
+                      ],
+                    ),
+                  ),
+                )
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
