+++ lib/ui/pages/pages.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+import 'package:nutrisi_app/ui/shared/text_style.dart';
+
+part 'item_food.dart';
+part 'home_page.dart';
